Create Dice Rolling Generator.py
This programme simulates rolling one or two dice, and it's one of the most approachable Python projects for people who are new to coding. It's also an excellent method to reinforce your knowledge of loops, conditional expressions, and user-defined functions. Whether you learn them from an online course or a Python book, these are essential abilities for anyone new to Python and will probably be some of the first things you learn.

It's a reasonably straightforward programme that mimics the randomness of rolling dice using the Python random module. It's one of our basic Python projects. You'll also notice that once you roll the dice, the screen is cleared using the OS module.

Keep in mind that you can imitate polyhedral dice, which are frequently used in board and roleplaying games, by setting the maximum dice value to any number.